Fix xclip hanging if fb output is piped elsewhere
xclip forks a child that listens for clipboard requests and returns the
copied content when necessary. When used with a pipe (`echo test | fb |
cat`) this leads to cat blocking because cat's stdin is still open.

I'm not sure why that happens, but closing stdout and stderr for xclip
fixes the problem. This has also been done like this in the older shell
client.

Signed-off-by: Florian Pritz <bluewind@xinu.at>