Fix error when repasting a single URL/ID
`fb id://1234 --name test` would skip uploading the file since it is
already uploaded. With this change, the file is first downloaded and
then uploaded again to result in a new ID.

Setting the file name via --name does not work yet for this example
command and is not addressed by this commit.

Signed-off-by: Florian Pritz <bluewind@xinu.at>